DVT-#128 모집 상세 페이지 레이아웃 변경
- 댓글 추가에 따른 변경
- 본문 글자색 변경
- 타이틀에 이모지 추가
- 게시글 등록날짜 시간 안보이게 처리
- 게시글 수정 버튼 및 수정 완료 버튼 추가하여 수정 버튼을 눌러야 수정할 수 있게 처리

+++ src/components/GatherDetail/GatherDetailContainer.tsx
@@ -1,16 +1,10 @@
-// TODO: API 연동 후 콘솔창은 지울 예정이다.
-/* eslint-disable no-console */
-
 import { useParams } from "react-router-dom";
 import useApplyGather from "../../hooks/useApplyGather";
 import useCancelAppliedGather from "../../hooks/useCancelApplyGather";
 import useGetGatherDetail from "../../hooks/useGetGatherDetail";
-
+import { common } from "../../constants";
 import GatherDetail from "./GatherDetail";
-import useCreateGatherComment from "../../hooks/useCreateGatherComment";
 import useDeleteGather from "../../hooks/useDeleteGather";
-import useDeleteComment from "../../hooks/useDeleteComment";
-import useEditComment from "../../hooks/useEditComment";
 import useEditGatherDetail from "../../hooks/useEditGatherDetail";
 import useClosGather from "../../hooks/useCloseGather";
 
@@ -19,57 +13,37 @@ const GatherDetailContainer = () => {
 
   const { data } = useGetGatherDetail(id);
 
-  console.log(data);
-
   const { cancelAppliedGather } = useCancelAppliedGather();
   const { applyGather } = useApplyGather();
-  const { createComment } = useCreateGatherComment();
   const { deleteGather } = useDeleteGather();
-  const { deleteComment } = useDeleteComment();
-  const { editComment } = useEditComment();
   const { editGather } = useEditGatherDetail();
   const { closeGather } = useClosGather();
 
-  // TODO: API 호출 로직 작성 필요
-
   const handleGatherDelete = (gatherId) => {
-    console.log("모임 삭제", gatherId);
-    deleteGather(gatherId);
+    // eslint-disable-next-line no-alert
+    // eslint-disable-next-line no-restricted-globals
+    if (confirm(common.message.CONFIRM_DELETE)) {
+      deleteGather(gatherId);
+    }
   };
 
   const handleGatherClose = (gatherId) => {
-    console.log("모임 마감", gatherId);
-    closeGather(gatherId);
+    // eslint-disable-next-line no-alert
+    // eslint-disable-next-line no-restricted-globals
+    if (confirm(common.message.CONFIRM_MODIFY)) {
+      closeGather(gatherId);
+    }
   };
 
   const handleGatherApply = (gatherId) => {
     applyGather(gatherId);
-    console.log("모임 신청", gatherId);
   };
 
   const handleGatherCancel = (gatherId) => {
     cancelAppliedGather(gatherId);
-
-    console.log("모임 취소", gatherId);
-  };
-
-  const handleCommentSubmit = (submitValue) => {
-    console.log("댓글 등록", submitValue);
-    createComment(submitValue);
-  };
-
-  const handleCommentDelete = (deleteValue) => {
-    console.log("댓글 삭제", deleteValue);
-    deleteComment(deleteValue);
-  };
-
-  const handleCommentEdit = (editValue) => {
-    console.log("댓글 수정", editValue);
-    editComment(editValue);
   };
 
   const handleGatherDetailEdit = (editValue) => {
-    console.log("모집 게시글 수정", editValue);
     editGather(editValue);
   };
 
@@ -78,16 +52,13 @@ const GatherDetailContainer = () => {
       {data ? (
         <GatherDetail
           gatherData={data?.data}
-          handleCommentDelete={handleCommentDelete}
-          handleCommentEdit={handleCommentEdit}
           handleGatherApply={handleGatherApply}
-          handleCommentSubmit={handleCommentSubmit}
           handleGatherCancel={handleGatherCancel}
           handleGatherClose={handleGatherClose}
           handleGatherDelete={handleGatherDelete}
           handleGatherDetailEdit={handleGatherDetailEdit}
         />
-      ) : undefined}
+      ) : null}
     </div>
   );
 };
